move server data module
