+++ src/server/utils.ts
@@ -1,0 +1,11 @@
+import http from 'node:http'
+
+export function parseBody(req: http.IncomingMessage): Promise<string> {
+  return new Promise(resolve => {
+    let body = ''
+    req.on('data', chunk => {
+      body += chunk
+    })
+    req.on('end', () => resolve(body))
+  })
+}
